detail.jsp, taskcontroller, view.jsp 에 login한 사용자가 테스크 삭제및수정 기능 등등 구현 +)add task member

+++ src/main/java/com/github/sbshin92/project_cal/controller/TaskController.java
@@ -135,19 +135,42 @@ public class TaskController {
 	    //테스크 삭제 
 	    @PostMapping("/deleteTask/{taskId}")
 	    public String deleteTask(@PathVariable int taskId, 
-	    						 @RequestParam int projectId) {
-	    	taskService.deleteTask(taskId);      
+	    						 @RequestParam int projectId,
+	    						 @RequestParam int userId,
+	    						 HttpSession httpsession) {
+	    	
+	        //로그인한 사용자가 이 task를 만든 사용자인지 체크 위함 
+	        //1) 테스크를 생성한 사용자의 아이디 필요
+	    	// userId
+	   	 	//2) 로그인한 사용자의 아이디 필요 
+	        UserVO userVO = (UserVO)httpsession.getAttribute("authUser");
+	        
+	         if ( userId == userVO.getUserId()) {
+	        	 taskService.deleteTask(taskId);
+	 	        return "redirect:/project/" + projectId; 
+
+	         } else {
+	        	 return "error/404";
+	         }
 	       
-	        return "redirect:/project/" + projectId; 
 	    }
 
 	    //테스크 수정 
 	    @PostMapping("/updateTask/{taskId}")
-	    public String updateTask(@PathVariable int taskId, @ModelAttribute TaskVO taskVO) {
-	        taskVO.setTaskId(taskId);
-	        taskService.updateTask(taskVO);
-	  
-	        return "redirect:/project/" + taskVO.getProjectId(); 
+	    public String updateTask(@PathVariable int taskId, @ModelAttribute TaskVO taskVO ,HttpSession session) {
+	       
+	    	//로그인한 사용자가 이 task를 만든 사용자인지 체크 후 수정
+	        UserVO userVO = (UserVO)session.getAttribute("authUser");
+
+	        if (taskVO.getUserId() == userVO.getUserId()) {
+	        	taskVO.setTaskId(taskId);
+		        taskService.updateTask(taskVO);
+		        return "redirect:/project/" + taskVO.getProjectId(); 
+	        } else {
+	        	return "error/404"; 
+	        }
+	    	
+	    	
 	    }	    
 	    
 	    //해당 테스크 상세 페이지 조회 안에 
@@ -155,6 +178,8 @@ public class TaskController {
 	    @GetMapping("/viewTask/{taskId}")
 	    public String viewTask(@PathVariable int taskId, Model model) {
 	        TaskVO task = taskService.findById(taskId);
+	        
+
 	        
 	        // UserTasks에 있는 멤버 조회
 	        List<UsersTasksVO> usersTasksVO = taskService.getUserTasksMember(taskId); 
@@ -166,13 +191,28 @@ public class TaskController {
 
 	    //해당 테스크에 멤버 추가 
 	    @PostMapping("/members/{taskId}")
-	    public String addMemberToTask(@PathVariable int taskId, @RequestParam int userId, @RequestParam int projectId) {
+	    public String addMemberToTask(@PathVariable int taskId, 
+    								@RequestParam int userId,
+    								@RequestParam int addUserId,
+    								@RequestParam int projectId, 
+    								HttpSession httpsession) {
 	        try { 
-	            taskService.addMemberToTask(userId, taskId, projectId);
+	        	//로그인한 사용자가 이 task를 만든 사용자인지 체크 후 추가
+		        UserVO userVO = (UserVO)httpsession.getAttribute("authUser");
+		    
+		        if (userId == userVO.getUserId()) {
+		        	taskService.addMemberToTask(addUserId, taskId, projectId);
+		        	return "redirect:/tasks/viewTask/" + String.valueOf(taskId);
+		        	
+		        } else {
+		        	 return "error/404";
+		        }
+
 	        } catch (Exception e) {
 	            e.printStackTrace();
-	        }
-	        return "redirect:/tasks/viewTask/" + String.valueOf(taskId);
+	        } 
+	        
+	        return "task/view";
 	    }
 	    
 	    //해당 테스크 멤버 삭제
